+++ spec/active_async/active_record_spec.rb
@@ -9,21 +9,20 @@ describe ActiveAsync::ActiveRecord do
       subject
     end
 
-    describe "included", :stub_resque do
-      it "should define async callback for save" do
-        Blog.should_receive(:expensive_save_method)
-        subject.save
-      end
+    it "should define async callback for save" do
+      Blog.should_receive(:expensive_save_method)
+      subject.save
+    end
 
-      it "should define async callback for update" do
-        Blog.should_receive(:expensive_update_method)
-        subject.update_attributes(:title => "foo")
-      end
+    it "should define async callback for update" do
+      Blog.should_receive(:expensive_update_method)
+      subject.update_attributes(:title => "foo")
+    end
 
-      it "should define async callback for create" do
-        Blog.should_receive(:expensive_create_method)
-        Blog.create
-      end
+    it "should define async callback for create" do
+      Blog.should_receive(:expensive_create_method)
+      Blog.create
     end
   end
+
 end