+++ spec/active_async/fake_queue_spec.rb
@@ -1,9 +1,9 @@
 require 'spec_helper'
 
-describe ActiveAsync::FakeResque do
+describe ActiveAsync::FakeQueue do
 
   it "should call perform with given klass and given arguments" do
     Post.should_receive(:perform).with(1, :expensive_method)
-    ActiveAsync::FakeResque.enqueue(Post, 1, :expensive_method)
+    ActiveAsync::FakeQueue.enqueue(Post, 1, :expensive_method)
   end
 end