+++ spec/active_async/rspec_spec.rb
@@ -5,18 +5,18 @@ describe "ActiveAsync::RSpec" do
   describe ":stub_resque block" do
 
     it "should set async mode to ::Resque in example", :stub_resque do
-      ActiveAsync.mode.should == :fake_resque
+      ActiveAsync.mode.should == :fake_queue
     end
 
     it "should stub_resque in example", :stub_resque do
-      ActiveAsync.background.should == ActiveAsync::FakeResque
+      ActiveAsync.background.should == ActiveAsync::FakeQueue
     end
 
-    it "should set async mode to ActiveAsync::FakeResque in example", :enable_resque do
+    it "should set async mode to ActiveAsync::FakeQueue in example", :enable_resque do
       ActiveAsync.mode.should == :resque
     end
 
-    it "should set async mode to ActiveAsync::FakeResque in example", :enable_resque do
+    it "should set async mode to ActiveAsync::FakeQueue in example", :enable_resque do
       ActiveAsync.background.should == ::Resque
     end
   end